refactor: relocate namespace configuration
